Promote state_reason to uint32_t in job_info_t.
This is already being passed around elsewhere as uint32_t in the
job_step_info_t, dbd_job_start_msg_t, and  acct_policy_validate_args_t
structs, and as an argument to acct_policy_validate().

(It's also "int unsigned not null" in the database which is uint32_t.)

+++ slurm/slurm.h
@@ -1860,7 +1860,7 @@ typedef struct job_info {
 				      *	either srun or the lowest slurmd version
 				      *	it is talking to */
 	char *state_desc;	/* optional details for state_reason */
-	uint16_t state_reason;	/* reason job still pending or failed, see
+	uint32_t state_reason;	/* reason job still pending or failed, see
 				 * slurm.h:enum job_state_reason */
 	char *std_err;		/* pathname of job's stderr file */
 	char *std_in;		/* pathname of job's stdin file */
